requirement.txt
autopep8==1.5.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
python3-openid==3.2.0
pytz==2020.5
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.2


